swapped .meta/example.rs with src/lib.rs properly
