+++ exercises/practice/pov/.meta/example.rs
@@ -1,35 +1,95 @@
-#[derive(Clone, PartialEq, Eq)]
-pub struct Tree {}
+use std::ops::{Deref, DerefMut};
 
-impl Tree {
+#[derive(Clone, PartialEq)]
+pub struct Tree<T: Clone + PartialEq> {
+    label: T,
+    children: Vec<Box<Tree<T>>>,
+}
+
+impl<T: Clone + PartialEq> Tree<T> {
     #[must_use]
-    pub fn new(_label: &str) -> Self {
-        todo!("Implement a function that creates a new Tree with given label");
+    pub fn new(label: T) -> Self {
+        Self {
+            label,
+            children: Default::default(),
+        }
     }
 
     #[must_use]
-    pub fn with_children(_label: &str, _children: &[Tree]) -> Self {
-        todo!("Implement a function that creates a new Tree with given label and Children");
+    pub fn with_children(label: T, children: Vec<Self>) -> Self {
+        Self {
+            label,
+            children: children.into_iter().map(Box::new).collect(),
+        }
     }
 
     #[must_use]
-    pub fn get_label(&self) -> String {
-        todo!("Implement getter for label.");
+    pub fn get_label(&self) -> T {
+        self.label.clone()
     }
 
     #[must_use]
     pub fn get_children(&self) -> Vec<&Self> {
-        todo!("Implement getter for children.");
+        self.children.iter().map(Box::deref).collect()
     }
 
     #[must_use]
-    pub fn pov_from(self, _from: &str) -> Option<Self> {
-        todo!("Implement a function that reparents Tree with 'from' as root.");
+    pub fn pov_from(&self, from: &T) -> Option<Self> {
+        // list of (child, parent, child's index in parent.children)
+        let mut lookup = vec![(self, None, None)];
+        let mut stack = vec![self];
+        while let Some(parent) = stack.pop() {
+            if &parent.label == from {
+                return self.reparent(parent, lookup.as_slice()).into();
+            }
+            lookup.extend(
+                parent
+                    .children
+                    .iter()
+                    .map(Box::deref)
+                    .enumerate()
+                    .map(|(i, child)| (child, Some(parent), Some(i))),
+            );
+            stack.extend(parent.children.iter().map(Box::deref));
+        }
+        None
+    }
+
+    // lookup is list of (child, parent, child's index in parent.children)
+    #[must_use]
+    fn reparent(&self, parent: &Self, lookup: &[(&Self, Option<&Self>, Option<usize>)]) -> Self {
+        let mut new_root = parent.clone();
+        let mut current = parent;
+        let mut clone_mut = &mut new_root;
+        let find_parent = |child| lookup.iter().find(|(c, _p, _i)| *c == child);
+        while let Some(&(_, Some(parent), Some(index))) = find_parent(current) {
+            let mut parent_clone = parent.clone();
+            parent_clone.children.swap_remove(index);
+            clone_mut.children.push(Box::new(parent_clone));
+            current = parent;
+            let new_box = clone_mut
+                .children
+                .last_mut()
+                .expect("We just inserted node, this is not empty");
+            clone_mut = new_box.deref_mut();
+        }
+        new_root
     }
 
     #[must_use]
-    pub fn path_to(self, _from: &str, _to: &str) -> Option<Vec<String>>
-    {
-        todo!("Implement a function that returns the list of labels in the shortest path from 'from' to 'to'");
+    pub fn path_to(&self, from: &T, to: &T) -> Option<Vec<T>> {
+        if from != &self.label {
+            return self.pov_from(from).and_then(|pov| pov.path_to(from, to));
+        }
+        if to == &self.label {
+            return Some(vec![self.label.clone()]);
+        }
+        for child in self.children.iter() {
+            if let Some(mut path) = child.path_to(&child.label, to) {
+                path.insert(0, self.label.clone());
+                return Some(path);
+            }
+        }
+        None
     }
 }
